reference consolidated `tile-layer` directory
