+++ public/tutorial/accessible/script/fullscreen.js
@@ -1,6 +1,6 @@
 import { createFullscreenMap } from '../../../script/leaflet-fullscreen.js';
 import { marker } from '../../../leaflet-adapter/marker.js';
-import { tileLayer } from '../../../leaflet-adapter/tile-layer.js';
+import { tileLayer } from '../../../leaflet-adapter/tile-layer/tile-layer.js';
 import {
   attributionOsm,
   urlTemplateOsm,
