+++ public/tutorial/custom-marker-icons/script/custom-marker-icons.js
@@ -2,7 +2,7 @@ import {
   createMap,
 } from '../../../leaflet-adapter-deprecated/leaflet-adapter-deprecated.js';
 import { marker } from '../../../leaflet-adapter/marker.js';
-import { tileLayer } from '../../../leaflet-adapter/tile-layer.js';
+import { tileLayer } from '../../../leaflet-adapter/tile-layer/tile-layer.js';
 import {
   attributionOsm,
   urlTemplateOsm,
