+++ public/tutorial/geojson/script/geojson.js
@@ -5,7 +5,7 @@ import { circleMarker } from '../../../leaflet-adapter/circle-marker.js';
 import {
   geoJson as leafletGeoJson,
 } from '../../../leaflet-adapter/geo-json.js';
-import { tileLayer } from '../../../leaflet-adapter/tile-layer.js';
+import { tileLayer } from '../../../leaflet-adapter/tile-layer/tile-layer.js';
 import {
   attributionOsm,
   urlTemplateOsm,
