+++ public/tutorial/wms/script/topo-osm.js
@@ -1,7 +1,7 @@
 import {
   createMap,
 } from '../../../leaflet-adapter-deprecated/leaflet-adapter-deprecated.js';
-import { tileLayerWms } from '../../../leaflet-adapter/tile-layer-wms.js';
+import { tileLayerWms } from '../../../leaflet-adapter/tile-layer/web-map-service.js';
 
 tileLayerWms({
   baseUrl: 'https://ows.mundialis.de/services/service?',
